Revert "Revert "restructure the files and update Item class""
