+++ classes/author_class.rb
