+++ classes/item.rb
@@ -1,8 +1,8 @@
 require 'date'
 
 class Item
-  attr_accessor :publish_date, :archived, :label
-  attr_reader :author
+  attr_accessor :publish_date, :archived
+  attr_reader :author, :label, :genre
 
   def initialize(publish_date, archived: false)
     @id = Random.rand(1...1000)
@@ -10,14 +10,24 @@ class Item
     @archived = archived
   end
 
+  def genre=(genre)
+    @genre = genre
+    genre.items.push(self) unless genre.items.include?(self)
+  end
+
+  def label=(label)
+    @label = label
+    label.items.push(self) unless label.items.include?(self)
+  end
+
   def author=(author)
     @author = author
     author.items.push(self) unless author.items.include?(self)
   end
 
-  def can_be_archived?(publish_date)
+  def can_be_archived?
     current_year = Date.today.year
-    published_year = @publish_date.to_i
+    published_year = Date.parse(@publish_date).year
     check_diff = current_year - published_year
     return false unless check_diff > 10
