+++ classes/musicalbum.rb
