+++ spec/genre_test.rb
@@ -1,6 +1,6 @@
 require 'rspec'
-require_relative 'genre'
-require_relative 'item'
+require_relative '../classes/genre'
+require_relative '../classes/item'
 
 describe Genre do
   before(:each) { @genre = Genre.new(88, 'Heavy metal') }
