+++ spec/item_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../item'
+require_relative '../classes/item'
 require 'rspec'
 
 describe Item do
